adding turnoController

+++ src/main/java/com/digitalhouse/clinicaOdontologica/controller/OdontologoController.java
@@ -2,6 +2,9 @@ package com.digitalhouse.clinicaOdontologica.controller;
 
 import com.digitalhouse.clinicaOdontologica.model.Odontologo;
 import com.digitalhouse.clinicaOdontologica.service.OdontologoService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,32 +20,45 @@ public class OdontologoController {
     }
 
     @GetMapping("/buscar/{id}")
-    public Odontologo buscarPorId(@PathVariable Integer id){
-        return odontologoService.getOdontologoById(id);
-
+    public ResponseEntity<?> buscarPorId(@PathVariable Integer id){
+        Odontologo odontologo = odontologoService.getOdontologoById(id);
+        if(odontologo != null){
+        return ResponseEntity.ok(odontologo);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("El odontologo no fue encontrado");
     }
 
     @GetMapping("/buscarTodos")
-    public List<Odontologo> buscarTodos(){
-        return odontologoService.getAll();
-
+    public ResponseEntity<?> buscarTodos(){
+        List<Odontologo> odontologos = odontologoService.getAll();
+        if(odontologos != null){
+            return ResponseEntity.ok(odontologos);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("No se encontraron odontologos");
     }
 
     @PostMapping("/agregar")
-    public Odontologo agregarOdontologo(@RequestBody Odontologo odontologo){
-        return odontologoService.saveOdontologo(odontologo);
+    public ResponseEntity<Odontologo>agregarOdontologo(@RequestBody Odontologo odontologo){
+        return ResponseEntity.ok(odontologoService.saveOdontologo(odontologo));
     }
 
     @PutMapping("/modificar")
-    public String modificarOdontologo(@RequestBody Odontologo odontologo){
+    public ResponseEntity<String> modificarOdontologo(@RequestBody Odontologo odontologo){
+        Odontologo odontologoAModificar = odontologoService.getOdontologoById(odontologo.getId());
+        if(odontologoAModificar != null){
         odontologoService.updateOdontologo(odontologo);
-        return "El Odontologo fue modificado";
+        return ResponseEntity.status(HttpStatus.OK).body("El odontologo fue actualizado");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("El Odontologo no fue encontrado para modificarlo");
     }
 
     @DeleteMapping("/eliminar/{id}")
-    public String eliminarOdontologo(@PathVariable Integer id){
-        odontologoService.deleteOdontologo(id);
-        return "El Odontologo fue eliminado";
+    public ResponseEntity<?> eliminarOdontologo(@PathVariable Integer id){
+        Odontologo odontogoAEliminar = odontologoService.getOdontologoById(id);
+        if(odontogoAEliminar != null){
+            odontologoService.deleteOdontologo(id);
+            return ResponseEntity.status(HttpStatus.OK).body("El odontologo fue eliminado");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("El Odontologo no fue encontrado para eliminarlo");
     }
-
 }
